ai-server/requirements.txt
alembic==1.16.5
amazon-textract-textractor==1.9.2
boto3==1.40.61
botocore==1.40.61
fastapi==0.121.1
langgraph==0.6.11
langchain-community==0.3.31
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langchain-pinecone==0.2.13
psutil==7.1.3
psycopg2-binary==2.9.11
pydantic==2.12.4
google-genai==1.47.0
PyPDF2==3.0.1
pdfplumber==0.11.8
python-docx==1.2.0
python-dotenv==1.2.1
python-magic==0.4.27
python-multipart==0.0.20
python-pptx==1.0.2
requests==2.32.5
s3fs==2025.10.0
openai==2.7.2
SQLAlchemy==2.0.44
uvicorn==0.38.0
PyMuPDF==1.26.5
nltk==3.9.2
typing_extensions==4.15.0
PyJWT==2.10.1
